Split SparkSubmissionContentPanel from SparkSubmissionExDialog so the Spark submission UI can be reused
